Immer autofreezes objects. When the object is really large, this gets more and more costly. I switched us to mutative, which does not rely on auto-freeze. [They have a comparison about immer vs mutative here](https://mutative.js.org/docs/extra-topics/comparison-with-immer). 

I made a bunch of transactions in our perf tester. 

**Before** 
[many_tx_immer.json](https://github.com/user-attachments/files/17065226/many_tx_immer.json)

A single transaction ontop of the bench query on immer is about 8ms. As you add more pending transactions, it grows to about 40ms.
![image](https://github.com/user-attachments/assets/5d5239ae-b2e6-479a-b6d1-56b8c86d7a92)

**After** 
[many_tx_mutative.json](https://github.com/user-attachments/files/17065227/many_tx_mutative.json)

Mutative is about 3ms on a single transaction. As I add more pending transactions, the size _seems to stay_ at 3ms! 

<img width="498" alt="CleanShot 2024-09-19 at 12 10 26@2x" src="https://github.com/user-attachments/assets/f2845090-51cb-45b2-9128-0ff744d189d4">

@dwwoelfel @markyfyi @nezaj